project/requirements.txt
certifi==2021.10.8
Django==1.8.5
gunicorn==19.10.0
psycopg2-binary==2.8.6
python-dotenv==0.18.0
typing==3.10.0.0
